backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dateutil==2.8.2


